requirements.txt
python-dotenv
playwright
sentence-transformers
spacy
fastapi
uvicorn
apscheduler
requests
jinja2
sqlalchemy>=2.0
psycopg2-binary
alembic
sqlmodel